Bug fix on phase locking
Previously, it was possible to get locked in a phase where there were
no actions to take but you were required to take an action.  For
example, if there were no units to deploy, you may still be expected to
deploy units.

Turn order for phases now checks how many units can act in a given phase
before generating the turn order for that phase.