Debugging for scenario board creation.
When creating the board from a scenario that has fixed boards,
the offsets do not get applied correctly. This is because in the
random version, we loop over the offsets but here we loop over the
boards themselves. To fix this, I will need to replace this map with a
loop or replace the structure they are mapping over with a map that has
the data.