Github workflow changes before 0.2 release
Because of the need to split resources and data folders, we cannot use
jPackage yet because I don't know how to include files in a deb file
(or an exe/msi for that matter). Therefore we're taking the MegaMek
approach of including a startup script that will launch the jar file.