Add support for facing.
No support yet for changing facing. Facing defaults N unless changed by
the scenario file and cannot be changed.

Also addresses a "bug" in scenario parsing which failed to interpret the
"(Standard)" tag found at the end of some Clan units as referring to
the "Prime" or model-less version of the unit.

Eventually, get-unit will become the standard way that I get a unit from
the MUL, accounting for issues with Battle Armor, etc. But for now, it
only serves to address this specific issue.