Pull Attacks into their own file.
In order to facilitate better reporting, attacks needed to be rewritten
to carry metadata for the attack. This was a good opportunity to add
reporting and details to the attack as well for use in the round reports
and the game logs. Attacks now live in the attack.clj file. There are
still some elements related to damage which are in combat_unit.clj, but
they will be migrated soon.