Split Combat units by functionality.
There is a lot of code related to Combat Units, but that code has been
grouped all in the complicated Combat Unit file. This resulted in a
large, complicated file that was hard to navigate. Now it is split
between combat_unit.clj, movement.clj, and attacks.clj. The
combat_unit.clj file is responsible for creating combat-units and for
organizing the "getters" and "setters" for combat-units.

Attacks is responsible for all logic related to attacks and damage.

Movement is responsible for all logic related to movement.