Current final iteration on round reports.
The only place I really need complex parsing for round reports is for
attacks. So I have simply had my attack function return all the data
required to write the log to the event, which then passes it off to a
parsing function to write up for the round report.