Add :changes map to combat-units
This allows us to save changes during phases so that if a unit takes
damage, it doesn't take effect until the end of the round.

Next step is to apply the changes to the units at the end of each round.
Then we need to incorporate the partial damage into the display.