Pull business logic out of GUI Code
In discussing with Perplexity how to handle AI player turns, I
discovered a way to pull logic OUT of my events so that they don't have
to collect a bunch of data and perform a bunch of operations on that
data. The ultimate goal is that they simply pass the *entire state atom*
into functions living mostly in the turn_manager file which will then
handle all the state manipulation.