Rewrite to switch to spec-schemaed maps
This will also make use of the tests again to give myself mroe
confidence.

This was a substantial change, but all these lines of code touch on
movement. I will now begin to work my way through the remaining files
alphabestically to switch them to spec and schemas, then I will work
through the GUI and finally the AI code