Change attack data structure to be changes only
Previously, we were passing around the entire data structure (which
included the full targeting information for recalculating the to-hit and
two complete copies of the attacker and the defender to represent their
state before and after the attack). This change replaces that with only
the relevant data required to generate a combat report and the changes
to the units.

Most significantly, this represents the changes to the units as vector
of vectors. The first element is the "path" through the state atom to
get to the value to change and the second element is the new value. This
is combined with a new function in turn_manager which will take a
game-state and a vector like the one described above and apply all those
changes to the game-state, returning the new game-state.